fixing bug with on-hold directive
Adding the on-hold directive caused a bug on the actual device where the
edit page would show behind the pop up to delete. Added an on-tap to
distinguish between event types. The implementation of on-tap handler
needed to be crude and use a window.location.href call due to a bug in
Angular's UI router where parameters are lost in transition when using
state.go